app/requirements.txt
# Server and Framework
fastapi
uvicorn
python-dotenv

# AI and LangChain
langchain
langchain-community
langchain-core
langchain-huggingface
transformers
torch
accelerate

# Database
langchain-mongodb
pymongo
motor

# Git Integration
GitPython
langchain-text-splitters